requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
openai-whisper>=20231117
yt_dlp>=2023.11.16
newspaper3k>=0.2.8
readability-lxml>=0.8.1
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
